+++ app/(futureco)/api/edge-geolocation/route.ts
