+++ app/(futureco)/blog/[slug]/page.tsx
