+++ app/(futureco)/blog/page.tsx
