+++ app/(futureco)/carburants/prix-a-la-pompe/rules.yaml
