+++ app/(futureco)/documentation/[...dottedName]/page.tsx
