+++ app/(futureco)/fin/[diapo]/page.tsx
