+++ app/(futureco)/montest/[...mavar]/page.tsx
