+++ app/(futureco)/scenarios/page.tsx
