+++ app/(futureco)/simulateur/[...dottedName]/Simulateur.tsx
