+++ app/(futureco)/simulateur/[...dottedName]/SimulateurContent.tsx
