+++ app/(futureco)/simulateur/[...dottedName]/configBuilder.ts
