+++ app/(futureco)/simulateur/[...dottedName]/error.tsx
