+++ app/(futureco)/simulateur/[...dottedName]/page.tsx
