+++ app/(futureco)/wiki/Wiki.tsx
