+++ app/(futureco)/wiki/page.tsx
