+++ app/voyage/robots.txt
@@ -1,0 +1,4 @@
+User-Agent: *
+Allow: /
+
+Sitemap: https://futur.eco/sitemap.xml
