+++ providers/Providers.tsx
@@ -1,4 +1,3 @@
-import { ThemeColorsProvider } from '@/components/utils/colors'
 import { PropsWithChildren } from 'react'
 import ReduxProvider from './ReduxProvider'
 
@@ -7,8 +6,6 @@ type P = PropsWithChildren
 export default function Providers({ children }: P) {
 	return (
 		// you can have multiple client side providers wrapped, in this case I am also using NextUIProvider
-		<ReduxProvider>
-			<ThemeColorsProvider>{children}</ThemeColorsProvider>
-		</ReduxProvider>
+		<ReduxProvider>{children}</ReduxProvider>
 	)
 }
